Ensure that DecimalField is not coerced to FloatField.

+++ tests/db_functions/math/test_acos.py
@@ -13,10 +13,10 @@ class ACosTests(TestCase):
     def test_decimal(self):
         DecimalModel.objects.create(n1=Decimal('-0.9'), n2=Decimal('0.6'))
         obj = DecimalModel.objects.annotate(n1_acos=ACos('n1'), n2_acos=ACos('n2')).first()
-        self.assertIsInstance(obj.n1_acos, float)
-        self.assertIsInstance(obj.n2_acos, float)
-        self.assertAlmostEqual(obj.n1_acos, math.acos(obj.n1))
-        self.assertAlmostEqual(obj.n2_acos, math.acos(obj.n2))
+        self.assertIsInstance(obj.n1_acos, Decimal)
+        self.assertIsInstance(obj.n2_acos, Decimal)
+        self.assertAlmostEqual(obj.n1_acos, Decimal(math.acos(obj.n1)))
+        self.assertAlmostEqual(obj.n2_acos, Decimal(math.acos(obj.n2)))
 
     def test_float(self):
         FloatModel.objects.create(f1=-0.5, f2=0.33)
