+++ tests/db_functions/math/test_atan.py
@@ -13,10 +13,10 @@ class ATanTests(TestCase):
     def test_decimal(self):
         DecimalModel.objects.create(n1=Decimal('-12.9'), n2=Decimal('0.6'))
         obj = DecimalModel.objects.annotate(n1_atan=ATan('n1'), n2_atan=ATan('n2')).first()
-        self.assertIsInstance(obj.n1_atan, float)
-        self.assertIsInstance(obj.n2_atan, float)
-        self.assertAlmostEqual(obj.n1_atan, math.atan(obj.n1))
-        self.assertAlmostEqual(obj.n2_atan, math.atan(obj.n2))
+        self.assertIsInstance(obj.n1_atan, Decimal)
+        self.assertIsInstance(obj.n2_atan, Decimal)
+        self.assertAlmostEqual(obj.n1_atan, Decimal(math.atan(obj.n1)))
+        self.assertAlmostEqual(obj.n2_atan, Decimal(math.atan(obj.n2)))
 
     def test_float(self):
         FloatModel.objects.create(f1=-27.5, f2=0.33)
