+++ tests/db_functions/math/test_atan2.py
@@ -12,8 +12,8 @@ class ATan2Tests(TestCase):
     def test_decimal(self):
         DecimalModel.objects.create(n1=Decimal('-9.9'), n2=Decimal('4.6'))
         obj = DecimalModel.objects.annotate(n_atan2=ATan2('n1', 'n2')).first()
-        self.assertIsInstance(obj.n_atan2, float)
-        self.assertAlmostEqual(obj.n_atan2, math.atan2(obj.n1, obj.n2))
+        self.assertIsInstance(obj.n_atan2, Decimal)
+        self.assertAlmostEqual(obj.n_atan2, Decimal(math.atan2(obj.n1, obj.n2)))
 
     def test_float(self):
         FloatModel.objects.create(f1=-25, f2=0.33)
