+++ tests/db_functions/math/test_cos.py
@@ -13,10 +13,10 @@ class CosTests(TestCase):
     def test_decimal(self):
         DecimalModel.objects.create(n1=Decimal('-12.9'), n2=Decimal('0.6'))
         obj = DecimalModel.objects.annotate(n1_cos=Cos('n1'), n2_cos=Cos('n2')).first()
-        self.assertIsInstance(obj.n1_cos, float)
-        self.assertIsInstance(obj.n2_cos, float)
-        self.assertAlmostEqual(obj.n1_cos, math.cos(obj.n1))
-        self.assertAlmostEqual(obj.n2_cos, math.cos(obj.n2))
+        self.assertIsInstance(obj.n1_cos, Decimal)
+        self.assertIsInstance(obj.n2_cos, Decimal)
+        self.assertAlmostEqual(obj.n1_cos, Decimal(math.cos(obj.n1)))
+        self.assertAlmostEqual(obj.n2_cos, Decimal(math.cos(obj.n2)))
 
     def test_float(self):
         FloatModel.objects.create(f1=-27.5, f2=0.33)
