+++ tests/db_functions/math/test_exp.py
@@ -13,10 +13,10 @@ class ExpTests(TestCase):
     def test_decimal(self):
         DecimalModel.objects.create(n1=Decimal('-12.9'), n2=Decimal('0.6'))
         obj = DecimalModel.objects.annotate(n1_exp=Exp('n1'), n2_exp=Exp('n2')).first()
-        self.assertIsInstance(obj.n1_exp, float)
-        self.assertIsInstance(obj.n2_exp, float)
-        self.assertAlmostEqual(obj.n1_exp, math.exp(obj.n1))
-        self.assertAlmostEqual(obj.n2_exp, math.exp(obj.n2))
+        self.assertIsInstance(obj.n1_exp, Decimal)
+        self.assertIsInstance(obj.n2_exp, Decimal)
+        self.assertAlmostEqual(obj.n1_exp, Decimal(math.exp(obj.n1)))
+        self.assertAlmostEqual(obj.n2_exp, Decimal(math.exp(obj.n2)))
 
     def test_float(self):
         FloatModel.objects.create(f1=-27.5, f2=0.33)
