+++ tests/db_functions/math/test_log.py
@@ -12,8 +12,8 @@ class LogTests(TestCase):
     def test_decimal(self):
         DecimalModel.objects.create(n1=Decimal('12.9'), n2=Decimal('3.6'))
         obj = DecimalModel.objects.annotate(n_log=Log('n1', 'n2')).first()
-        self.assertIsInstance(obj.n_log, float)
-        self.assertAlmostEqual(obj.n_log, math.log(obj.n2, obj.n1))
+        self.assertIsInstance(obj.n_log, Decimal)
+        self.assertAlmostEqual(obj.n_log, Decimal(math.log(obj.n2, obj.n1)))
 
     def test_float(self):
         FloatModel.objects.create(f1=2.0, f2=4.0)
