+++ tests/db_functions/math/test_mod.py
@@ -12,8 +12,8 @@ class ModTests(TestCase):
     def test_decimal(self):
         DecimalModel.objects.create(n1=Decimal('-9.9'), n2=Decimal('4.6'))
         obj = DecimalModel.objects.annotate(n_mod=Mod('n1', 'n2')).first()
-        self.assertIsInstance(obj.n_mod, float)
-        self.assertAlmostEqual(obj.n_mod, math.fmod(obj.n1, obj.n2))
+        self.assertIsInstance(obj.n_mod, Decimal)
+        self.assertAlmostEqual(obj.n_mod, Decimal(math.fmod(obj.n1, obj.n2)))
 
     def test_float(self):
         FloatModel.objects.create(f1=-25, f2=0.33)
