+++ tests/db_functions/math/test_sin.py
@@ -13,10 +13,10 @@ class SinTests(TestCase):
     def test_decimal(self):
         DecimalModel.objects.create(n1=Decimal('-12.9'), n2=Decimal('0.6'))
         obj = DecimalModel.objects.annotate(n1_sin=Sin('n1'), n2_sin=Sin('n2')).first()
-        self.assertIsInstance(obj.n1_sin, float)
-        self.assertIsInstance(obj.n2_sin, float)
-        self.assertAlmostEqual(obj.n1_sin, math.sin(obj.n1))
-        self.assertAlmostEqual(obj.n2_sin, math.sin(obj.n2))
+        self.assertIsInstance(obj.n1_sin, Decimal)
+        self.assertIsInstance(obj.n2_sin, Decimal)
+        self.assertAlmostEqual(obj.n1_sin, Decimal(math.sin(obj.n1)))
+        self.assertAlmostEqual(obj.n2_sin, Decimal(math.sin(obj.n2)))
 
     def test_float(self):
         FloatModel.objects.create(f1=-27.5, f2=0.33)
