+++ tests/db_functions/math/test_sqrt.py
@@ -13,10 +13,10 @@ class SqrtTests(TestCase):
     def test_decimal(self):
         DecimalModel.objects.create(n1=Decimal('12.9'), n2=Decimal('0.6'))
         obj = DecimalModel.objects.annotate(n1_sqrt=Sqrt('n1'), n2_sqrt=Sqrt('n2')).first()
-        self.assertIsInstance(obj.n1_sqrt, float)
-        self.assertIsInstance(obj.n2_sqrt, float)
-        self.assertAlmostEqual(obj.n1_sqrt, math.sqrt(obj.n1))
-        self.assertAlmostEqual(obj.n2_sqrt, math.sqrt(obj.n2))
+        self.assertIsInstance(obj.n1_sqrt, Decimal)
+        self.assertIsInstance(obj.n2_sqrt, Decimal)
+        self.assertAlmostEqual(obj.n1_sqrt, Decimal(math.sqrt(obj.n1)))
+        self.assertAlmostEqual(obj.n2_sqrt, Decimal(math.sqrt(obj.n2)))
 
     def test_float(self):
         FloatModel.objects.create(f1=27.5, f2=0.33)
