+++ tests/db_functions/math/test_tan.py
@@ -13,10 +13,10 @@ class TanTests(TestCase):
     def test_decimal(self):
         DecimalModel.objects.create(n1=Decimal('-12.9'), n2=Decimal('0.6'))
         obj = DecimalModel.objects.annotate(n1_tan=Tan('n1'), n2_tan=Tan('n2')).first()
-        self.assertIsInstance(obj.n1_tan, float)
-        self.assertIsInstance(obj.n2_tan, float)
-        self.assertAlmostEqual(obj.n1_tan, math.tan(obj.n1))
-        self.assertAlmostEqual(obj.n2_tan, math.tan(obj.n2))
+        self.assertIsInstance(obj.n1_tan, Decimal)
+        self.assertIsInstance(obj.n2_tan, Decimal)
+        self.assertAlmostEqual(obj.n1_tan, Decimal(math.tan(obj.n1)))
+        self.assertAlmostEqual(obj.n2_tan, Decimal(math.tan(obj.n2)))
 
     def test_float(self):
         FloatModel.objects.create(f1=-27.5, f2=0.33)
